backend/requirements.txt

fastapi
psycopg2-binary
jsonschema
